requirements.txt
numpy
sounddevice
pystray
pillow